Making the ACFT Calculator
Implemented a basic ACFT Calculator However it only has females fully working right now. However i found a way to implement the scores function in another method that will enable me to update the score of the person based off the times or scores the enter interesting wording of choice there but in general that is what I have implemented so far. Next should be implementing the rest of the code for males with there respective scores and times and then adding the full functionality so I can output a score out of 600.